requirements.txt
Flask
gunicorn
openai==1.35.15
requests
python-dotenv
Pillow
